pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=1.46.0",
    "sift-stack-py>=0.9.1",
    "streamlit>=1.50.0",
    "supabase>=2.6.0",
]
